==> 1. Meldunki_download.ipynb <==
from bs4 import BeautifulSoup
import urllib3

from bs4 import BeautifulSoup
import urllib3
import pandas as pd
import PyPDF2 
import requests
import io
from io import BytesIO
from io import StringIO
import PyPDF2
from PyPDF2 import PdfFileReader
import re

df = pd.DataFrame([[0 for i in range(0,19)]], columns=["Year", "Range", "Amount", "Dolnośląskie", "Kujawsko-Pomorskie", "Lubelskie", "Lubuskie", "Łódzkie", "Małopolskie", "Mazowieckie", "Opolskie", "Podkarpackie", "Podlaskie", "Pomorskie", "Śląskie", "Świętokrzyskie", "Warmińsko-Mazurskie", "Wielkopolskie", "Zachodniopomorskie"])
            
for i in range(2010,2020+1):
    x = "http://wwwold.pzh.gov.pl/oldpage/epimeld/grypa/"+str(i)+"/"+str(i)+".htm"

    soup = BeautifulSoup(x, 'html.parser')

    http = urllib3.PoolManager()
    response = http.request('GET', x)
    soup = BeautifulSoup(response.data)
    for a in soup.find_all('a', href=True):
        if a['href'].startswith("G"):
            x = "http://wwwold.pzh.gov.pl/oldpage/epimeld/grypa/"+str(i)+"/"+a['href']

            r = requests.get(x)
            f = io.BytesIO(r.content)
            reader = PdfFileReader(f)
            contents = reader.getPage(0).extractText().split('\n')
            contents = (', '.join(map(str, contents)))
#             print(contents)

            try:
                start = re.search(r"POLSKA , ",contents).end()
                nr = contents[start:start+7]
                nr = nr.split(' ')
                
                start = re.search(r"Dolnoskie , ",contents).end()
                dolnoslaskie = contents[start:start+7]
                dolnoslaskie = dolnoslaskie.split(' ')
                
                start = re.search(r"Kujawsko-Pomorskie , ",contents).end()
                kujawskie = contents[start:start+7]
                kujawskie = kujawskie.split(' ')
                
                start = re.search(r"Lubelskie , ",contents).end()
                lubel = contents[start:start+7]
                lubel = lubel.split(' ')
                
                start = re.search(r" Lubuskie , ",contents).end()
                lubuskie = contents[start:start+7]
                lubuskie = lubuskie.split(' ')
                
                start = re.search(r" Łódzkie , ",contents).end()
                lodzkie = contents[start:start+7]
                lodzkie = lodzkie.split(' ')

                start = re.search(r" Małopolskie , ",contents).end()
                malo = contents[start:start+7]
                malo = malo.split(' ')

                start = re.search(r"Mazowieckie , ",contents).end()
                mazo = contents[start:start+7]
                mazo = mazo.split(' ')
                
                start = re.search(r"Opolskie , ",contents).end()
                opol = contents[start:start+7]
                opol = opol.split(' ')
            
                start = re.search(r"Podkarpackie , ",contents).end()
                podk = contents[start:start+7]
                podk = podk.split(' ')
                
                start = re.search(r"Podlaskie , ",contents).end()
                podla = contents[start:start+7]
                podla = podla.split(' ')
                
                start = re.search(r"Pomorskie , ",contents).end()
                pom = contents[start:start+7]
                pom = pom.split(' ')
                
                start = re.search(r" ie , ",contents).end()
                slaskie = contents[start:start+7]
                slaskie = slaskie.split(' ')
                
                start = re.search(r"witokrzyskie , ",contents).end()
                swieto = contents[start:start+7]
                swieto = swieto.split(' ')
                
                start = re.search(r" Warmisko-Mazurskie , ",contents).end()
                warmi = contents[start:start+7]
                warmi = warmi.split(' ')
                
                start = re.search(r" Wielkopolskie , ",contents).end()
                wielko = contents[start:start+7]
                wielko = wielko.split(' ')

                start = re.search(r"Zachodniopomorskie ",contents).end()
                zacho = contents[start:start+7]
                zacho = zacho.split(' ')

                data = i
                df_new = pd.DataFrame([[str(i), a['href'][5:8], nr[0], dolnoslaskie[0],\
                    kujawskie[0], lubel[0], lubuskie[0], lodzkie[0], malo[0], mazo[0],\
                    opol[0], podk[0], podla[0], pom[0], slaskie[0], swieto[0], warmi[0], wielko[0], zacho[0]]],\
                    columns=["Year", "Range", "Amount", "Dolnośląskie", "Kujawsko-Pomorskie", "Lubelskie", "Lubuskie", "Łódzkie", "Małopolskie", "Mazowieckie", "Opolskie", "Podkarpackie", "Podlaskie", "Pomorskie", "Śląskie", "Świętokrzyskie", "Warmińsko-Mazurskie", "Wielkopolskie", "Zachodniopomorskie"])
                df = df.append(df_new)
            except:
                pass
           

df.to_csv("2007-2020.csv", index=False)
df

df = pd.read_csv("2007-2020.csv")
df = df[1:]

df["Date"] = [0 for i in range(len(df))]
Date_column = []

#Add correct type of date
for i in range(len(df)):
    m = df.Range[i+1][0:2]
    d = df.Range[i+1][2:3]
    if d == "A":
        d = "01"
    elif d == "B":
        d = "08"
    elif d == "C":
        d = "16"
    elif d == "D":
        d = "23"
    Date = str(df.Year[i+1])+m+d
    Date_column.append(Date)
df = df.drop(columns="Range")
    
df.Date = Date_column 
df.Date = pd.to_datetime(df.Date, format='%Y%m%d', errors='ignore') 
cols = df.columns.tolist()
# df=df[cols[-1:]+cols[:-1]]
df
df.to_csv('2007-2020.csv', index=False)

df_meld = pd.read_csv("2007-2020.csv")
df_meld.head()

df_new = pd.DataFrame([[0 for i in range(0,3)]], columns=["Date", "Place", "Ill"])

for i in range(3,19): #for each voideship
    for j in range(df_meld.shape[0]): #for each date
        Date = df_meld.Date[j]
        Place = df_meld.columns[i]
        Ill = df_meld.iloc[j:j+1,i:i+1].values[0][0]
        df_new2 = pd.DataFrame([[Date, Place, Ill]],  columns=["Date", "Place", "Ill"])
        df_new = pd.concat([df_new, df_new2])

df_new = df_new[1:]
df_new.to_csv("2007-2020.csv", index=False)

df_meld = pd.read_csv("2007-2020.csv")
df_meld.head()

df_k = pd.read_csv("Klimat_d_t.csv", header=0)
df_k['Date'] = df_k['Rok'].astype(str).str.cat(df_k['Miesiąc'].astype(str),sep="-").astype(str).str.cat(df_k['Dzień'].astype(str),sep="-")
cols = df_k.columns.tolist()
cols = cols[-1:] + cols[:-1]
df_k = df_k[cols]
df_k['Nazwa stacji'] = df_k['Nazwa stacji'].str.lower()
df_k = df_k[df_k.Date > '2010-01-01']

df = pd.DataFrame([[0 for i in range(0,19)]], columns=["Year", "Range", "Amount", "Dolnośląskie", "Kujawsko-Pomorski", "Lubelskie", "Lubuskie", "Łódzkie", "Małopolskie", "Mazowieckie", "Opolskie", "Podkarpackie", "Podlaskie", "Pomorskie", "Śląskie", "Świętokrzyskie", "Warmińsko-Mazurskie", "Wielkopolskie", "Zachodniopomorskie"])
 
#Second range - different schema
for i in range(2000,2006):
    x = "http://wwwold.pzh.gov.pl/oldpage/epimeld/grypa/"+str(i)+"/"+str(i)+".htm"

    soup = BeautifulSoup(x, 'html.parser')

    http = urllib3.PoolManager()
    response = http.request('GET', x)
    soup = BeautifulSoup(response.data)
    for a in soup.find_all('a', href=True):
        if a['href'].startswith("G"):
            x = "http://wwwold.pzh.gov.pl/oldpage/epimeld/grypa/"+str(i)+"/"+a['href']
            #Download
#Open downloaded and extract 

                

import PyPDF2 
  
pdfFileObj = open('G_20_01B.pdf', 'rb') 
pdfReader = PyPDF2.PdfFileReader(pdfFileObj) 
  
print(pdfReader.numPages) 
pageObj = pdfReader.getPage(0) 
  
# print(pageObj.extractText())
