==> 2. Klimat_preproc.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
%matplotlib inline
import datetime
from datetime import datetime
import numpy as np
import matplotlib as mpl

df_k = pd.read_csv("Klimat 2010.csv", header=0, parse_dates=['Date'])
df_k.head()

df_k = pd.read_csv("Klimat 2010.csv", header=0, parse_dates=['Date'])
df_k['Wojewodztwo'] = [0 for i in range(len(df_k))]
df_k.head()

df_k.head()

df_k.iloc[0:1,12:13] 

cities = pd.read_csv("TERC_Urzedowy_2020-01-28.csv", header=0, sep=";")
cities.head()

cities = pd.read_csv("TERC_Urzedowy_2020-01-28.csv", header=0, sep=";")
cities.fillna(0,inplace=True)
cities.NAZWA = cities.NAZWA.str.lower()
dictionary = {}

for i in range(len(cities)):
    if cities.POW[i] == 0:
        dictionary[cities.WOJ[i]] = cities.NAZWA[i] 
cities_c = {}

for i in range(len(cities)):
    cities_c[cities.NAZWA[i]] = dictionary[cities.WOJ[i]]

for i in range(len(df_k)):
    try:
        df_k.iloc[i:i+1,11:12] = cities_c[df_k['Nazwa stacji'][i]]  #przypisanie do wojewodztw --> 
    except:
        df_k.iloc[i:i+1,11:12] = 0
        
df_k.to_csv("Klimat_woj.csv", index=False)
df_k.head()

df_k = pd.read_csv("Klimat_woj.csv", header=0, parse_dates=['Date'])
df_k.head()

df_k.Wojewodztwo[df_k['Nazwa stacji']=="inwałd"] = 'małopolskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="rabka"] = 'małopolskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="laliki"] = 'śląskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="obidowa"] = 'małopolskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="polana chochołowska"] = 'małopolskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="hala ornak"] = 'małopolskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="radziechowy"] = 'śląskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="ptaszkowa"] = 'małopolskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="łopuszna"] = 'małopolskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="mizerna"] = 'małopolskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="krościenko"] = 'małopolskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="niedzica"] = 'małopolskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="krynica"] = 'małopolskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="hala gąsienicowa"] = 'małopolskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="dolina pięciu stawów"] = 'małopolskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="morskie oko"] = 'małopolskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="piwniczna"] = 'małopolskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="biecz-grudna"] = 'małopolskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="wysowa"] = 'małopolskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="barwinek"] = 'podkarpackie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="sanok-trepcza"] = 'podkarpackie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="solina-jawor"] = 'podkarpackie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="baligród-mchawa"] = 'podkarpackie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="terka"] = 'podkarpackie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="jakuszyce"] = 'dolnośląskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="stuposiany"] = 'podkarpackie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="paprotki"] = 'dolnośląskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="pszenno"] = 'dolnośląskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="słoszów"] = 'dolnośląskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="długopole-zdrój"] = 'dolnośląskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="bolesławów"] = 'dolnośląskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="dobrogoszcz"] = 'pomorskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="stare olesno"] = 'opolskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="sukowice"] = 'opolskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="droniowice"] = 'śląskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="silniczka"] = 'łódzkie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="lgota górna"] = 'śląskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="olewin"] = 'małopolskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="bieruń stary"] = 'śląskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="kraków-obserwatorium"] = 'małopolskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="katowice pyrzowice"] = 'śląskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="sielec"] = 'lubelskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="borusowa"] = 'małopolskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="igołomia"] = 'małopolskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="jędrzejów-sudół"] = 'świętokrzyskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="święty krzyż"] = 'świętokrzyskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="chorzelów"] = 'podkarpackie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="zawada"] = 'dolnośląskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="ceber"] = 'świętokrzyskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="grabik"] = 'podkarpackie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="tomaszów bolesławiecki"] = 'dolnośląskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="polkowice dolne"] = 'dolnośląskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="chwałkowice"] = 'dolnośląskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="witaszyce"] = 'małopolskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="radzyń"] = 'wielkopolskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="smolice"] = 'wielkopolskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="grabownica"] = 'dolnośląskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="puczniew"] = 'łódzkie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="jarczew"] = 'lubelskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="bezek"] = 'lubelskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="krzyż"] = 'wielkopolskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="gorzyń"] = 'wielkopolskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="lubinicko-świebodzin"] = 'lubuskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="szamotuły-baborówko"] = 'wielkopolskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="paproć"] = 'wielkopolskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="kołuda wielka"] = 'kujawsko-pomorskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="warszawa-bielany"] = 'mazowieckie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="warszawa-filtry"] = 'mazowieckie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="warszawa-babice"] = 'mazowieckie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="warszawa-obserwatorium ii"] = 'mazowieckie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="cicibór"] = 'lubelskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="chrząstowo"] = 'pomorskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="radostowo"] = 'pomorskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="kasparus"] = 'pomorskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="dobrocin"] = 'warmińsko-mazurskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="biebrza-pieńczykówek"] = 'podlaskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="różanystok"] = 'podlaskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="karżniczka"] = 'pomorskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="borucino"] = 'wielkopolskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="gdańsk-rębiechowo"] = 'pomorskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="gdańsk-świbno"] = 'pomorskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="kmiecin"] = 'pomorskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="nowy cydzyn"] = 'podlaskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="bydgoszcz-szwederowo"] = 'kujawsko-pomorskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="marianowo ii"] = 'zachodniopomorskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="sobieszyn"] = 'lubelskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="głodowo"] = 'kujawsko-pomorskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="istebna-kubalonka"] = 'śląskie'
df_k.Wojewodztwo[df_k['Nazwa stacji']=="ząbkowice"] = 'dolnośląskie'



df_k.to_csv("Klimat_woj.csv", index=False)

[(df_k['Nazwa stacji'][df_k.Wojewodztwo==i]).unique() for i in uni] #jakie stacje dla jakiego woj

df_k[df_k.Wojewodztwo==0] #I wszystkie są przypisane

# df_k2.to_csv("df_k woj.csv", index=False)
df_k = pd.read_csv("Klimat_woj.csv", header=0, parse_dates=['Date'])
df_k.head()
