==> 4. DF_apply_preproces.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
%matplotlib inline
import datetime
from datetime import datetime
from datetime import timedelta
import numpy as np
import matplotlib as mpl
import re
from fastai.tabular import *

df_m = pd.read_csv("2007-2020.csv", header=0)
df_m.Place = df_m.Place.str.lower()
df_m.head() #(7680, 3)

df_t = pd.read_csv("trends.csv", header=0, names=["Token", "Place", "Date", "Value"], parse_dates=['Date'])
df_trends_grypa = df_t[df_t.Token=="grypa"]
df_trends_kaszel = df_t[df_t.Token=="kaszel"]
df_trends_grypa.head()

# df_m.to_csv("Meld_trends.csv", index=False)
df_m = pd.read_csv("Table.csv", header=0, parse_dates=['Date'])
df_m.head()

########ADD TRENDS_GRYPA
items_t = [i for i in df_trends_grypa.Date]

def func_t(date, woj, items):
    date = pd.to_datetime(date)
    data_trends = min(items, key=lambda x: abs(x - date)) #najbliższa data --> zbierz z niej wartości
    data_trends = str(data_trends)[0:10]
    try:
        value_trends = df_trends_grypa.Value[(df_trends_grypa.Date == data_trends)&
                                         (df_trends_grypa.Place==woj)].values[0]
    except:
        value_trends = 0
    return value_trends

df_m.head()

########ADD TRENDS_KASZEL
items_t = [i for i in df_trends_kaszel.Date]

def func_t_k(date, woj, items):
    date = pd.to_datetime(date)
    data_trends = min(items, key=lambda x: abs(x - date)) #najbliższa data --> zbierz z niej wartości
    data_trends = str(data_trends)[0:10]
    try:
        value_trends = df_trends_kaszel.Value[(df_trends_kaszel.Date == data_trends)&
                                         (df_trends_kaszel.Place==woj)].values[0]
    except:
        value_trends = 0
    return value_trends

df_k = pd.read_csv("Klimat_woj.csv", header=0, parse_dates=['Date'])
df_k.head()

df_o = pd.read_csv("Opad_woj.csv", header=0, parse_dates=['Date'])
df_o.head()

df_s = pd.read_csv("Syno_woj.csv", header=0, parse_dates=['Date'])
df_s.head()

df_s.iloc[1:2, 5:25]

df_m['TempDob'] = [0 for i in range(len(df_m))]
df_m['TempDob_7D'] = [0 for i in range(len(df_m))]
df_m['ZachmDob'] = [0 for i in range(len(df_m))]
df_m['Wiatr'] = [0 for i in range(len(df_m))]
df_m['Opad'] = [0 for i in range(len(df_m))]
df_m['DobZachm'] = [0 for i in range(len(df_m))]
df_m['Wiglot'] = [0 for i in range(len(df_m))]
df_m['Opad_7D'] = [0 for i in range(len(df_m))]

df_m.head()

df_m.iloc[0:1,3:4]

#Trends grypa
items_t = [i for i in df_trends_grypa.Date]

for i in range(len(df_m)):
    date = pd.to_datetime(df_m.iloc[i:i+1,0:1].values[0][0])
    data_trends = min(items_t, key=lambda x: abs(x - date)) #najbliższa data --> zbierz z niej wartości
    data_trends = str(data_trends)[0:10]
    df_m.iloc[i:i+1,3:4] = df_trends_grypa.Value[(df_trends_grypa.Date == data_trends)&
                            (df_trends_grypa.Place==df_m.iloc[i:i+1, 1:2].values[0][0])].values[0]
    print(i)
    

#Trends Kaszel
items_t = [i for i in df_trends_kaszel.Date]

for i in range(len(df_m)):
    date = pd.to_datetime(df_m.iloc[i:i+1,0:1].values[0][0])
    data_trends = min(items_t, key=lambda x: abs(x - date)) #najbliższa data --> zbierz z niej wartości
    data_trends = str(data_trends)[0:10]
    df_m.iloc[i:i+1,6:7] = df_trends_kaszel.Value[(df_trends_kaszel.Date == data_trends)&
                            (df_trends_kaszel.Place==df_m.iloc[i:i+1, 1:2].values[0][0])].values[0] 
    

#TempDob
for i in range(len(df_m)):
    df_m.iloc[i:i+1,4:5] =  np.mean(df_k['Średnia dobowa temperatura']
      [(df_k.Date==df_m.iloc[i:i+1, 0:1].values[0][0])&
       (df_k.Wojewodztwo==df_m.iloc[i:i+1, 1:2].values[0][0])])
    
df_m.head()

#TempDob_7D
for i in range(len(df_m)):
    date = pd.to_datetime(df_m.iloc[i:i+1,0:1].values[0][0])
    ranges = [date-timedelta(i) for i in range(0,7)]
    ranges = [str(ranges[i])[0:10] for i in range(0,7)]
    df_m.iloc[i:i+1,5:6] = np.mean(df_k['Średnia dobowa temperatura'][(df_k.Date.isin(ranges))&
       (df_k.Wojewodztwo==df_m.iloc[i:i+1, 1:2].values[0][0])])

df_m.iloc[i:i+1, 9:10]

#Wiatr-Dob
for i in range(len(df_m)):    
    df_m.iloc[i:i+1,7:8] = np.mean(df_k['Średnia dobowa prędkość wiatru']
    [(df_k.Date==df_m.iloc[i:i+1,0:1].values[0][0])
    &(df_k.Wojewodztwo==df_m.iloc[i:i+1, 1:2].values[0][0])])

#Opad
for i in range(len(df_m)):
    df_m.iloc[i:i+1,8:9] = np.mean(df_o['Suma dobowa opadów']
    [(df_o.Date==df_m.iloc[i:i+1,0:1].values[0][0])
    &(df_o.Wojewodztwo==df_m.iloc[i:i+1, 1:2].values[0][0])])

#Średnie dobowe zachmurzenie ogólne
for i in range(len(df_m)):
    df_m.iloc[i:i+1,9:10] = np.mean(df_s['Średnie dobowe zachmurzenie ogólne [oktanty]']
    [(df_s.Date==df_m.iloc[i:i+1,0:1].values[0][0])
    &(df_s.Wojewodztwo==df_m.iloc[i:i+1, 1:2].values[0][0])])

#Wilgot
for i in range(len(df_m)):
    df_m.iloc[i:i+1,10:11] = np.mean(df_s['Średnia dobowa wilgotność względna [%]']
    [(df_s.Date==df_m.iloc[i:i+1,0:1].values[0][0])
    &(df_s.Wojewodztwo==df_m.iloc[i:i+1, 1:2].values[0][0])])

#Opad 7D
for i in range(len(df_m)):
    date = pd.to_datetime(df_m.iloc[i:i+1,0:1].values[0][0])
    ranges = [date-timedelta(i) for i in range(0,7)]
    ranges = [str(ranges[i])[0:10] for i in range(0,7)]
    
    df_m.iloc[i:i+1,11:12] = np.mean(df_o['Suma dobowa opadów']
    [(df_o.Date.isin(ranges))&(df_o.Wojewodztwo==df_m.iloc[i:i+1, 1:2].values[0][0])])

df_m.head()

df_m.to_csv("Table.csv", index=False)
